requirements.txt
streamlit>=1.27
sqlalchemy
psycopg2-binary
python-dotenv